requirements.txt
opencv-contrib-python
Pillow
flask
requests
